Fix: Panoramas loading in iOS
Gallery.mm
Line 53 - date format needs to be set correctly
Line 110 - number of panoramas is limited to 10 (19 panoramas caused
the app to run out of memory)
Panorama is now saved as a jpeg and re-loaded in Unity. Cached file is
then removed.

AppController.cs
Lines 49&50 - added declaration of m_AutoVRModeSwap for iOS
		is this correct?

FileHandler.cs
Lines 154-164 - added function LoadLargeImage for iOS
		is this correct?

iOSUnityInterface.cs
Class is compiled to return dummy values if it is not compiled on iOS